+++ src/components/Programs/schoolProgramsTabs.ts
@@ -3,7 +3,7 @@ import Loader from "@/components/Loader.vue"
 import ProgressiveImage from "@/components/ProgressiveImage.vue"
 import DownloadLink from "@/components/DownloadLink.vue"
 import useSchoolPrograms from "@/factories/useSchoolPrograms"
-import useTabs from "@/factories/useTabs"
+import useTabs from "@/utils/useTabs"
 import { SchoolProgram } from "@/types/viewTypes"
 
 const SchoolProgramsTabs = defineComponent({
