+++ src/factories/useActivities.ts
@@ -2,7 +2,7 @@ import Vue from "vue"
 import VueCompositionApi, { computed } from "@vue/composition-api"
 import apiRoutes from "../../api/apiRoutes"
 import { Activity } from "@/types/viewTypes"
-import useAsyncData from "./useAsyncData"
+import useAsyncData from "../utils/useAsyncData"
 import { WpResponseData, WpImage } from "@/types/wordpressTypes"
 import AppUrls from "@/utils/urls"
 import { getWPTitle, getCustomField } from "@/utils/api"
