+++ src/factories/useNews.ts
@@ -3,7 +3,7 @@ import VueCompositionApi, { computed } from "@vue/composition-api"
 import apiRoutes from "../../api/apiRoutes"
 import { NewsPost, Activity } from "@/types/viewTypes"
 import helpers from "@/utils/customFields"
-import useAsyncData from "./useAsyncData"
+import useAsyncData from "../utils/useAsyncData"
 import { WpResponseData } from "@/types/wordpressTypes"
 import AppUrls from "@/utils/urls"
 import { ActivityType } from "@/types/customFieldsTypes"
