+++ src/factories/useProgramVideos.ts
@@ -2,7 +2,7 @@ import Vue from "vue"
 import VueCompositionApi, { Ref, computed } from "@vue/composition-api"
 import apiRoutes from "../../api/apiRoutes"
 import { ProgramVideo } from "@/types/viewTypes"
-import useAsyncData from "./useAsyncData"
+import useAsyncData from "../utils/useAsyncData"
 import { WpResponseData } from "@/types/wordpressTypes"
 import { getCustomField, getWPTitle } from "@/utils/api"
 import { ProgramVideoKeys } from "@/types/customFieldsKeysTypes"
