+++ src/factories/useSchoolPrograms.ts
@@ -2,11 +2,11 @@ import Vue from "vue"
 import VueCompositionApi, { computed } from "@vue/composition-api"
 import apiRoutes from "../../api/apiRoutes"
 import { SchoolProgram } from "@/types/viewTypes"
-import useAsyncData from "./useAsyncData"
+import useAsyncData from "../utils/useAsyncData"
 import { WpResponseData, WpImage } from "@/types/wordpressTypes"
 import { getCustomField, getWPTitle } from "@/utils/api"
 import { SchoolProgramKeys } from "@/types/customFieldsKeysTypes"
-import { Tab, Tabs } from "./useTabs"
+import { Tab, Tabs } from "../utils/useTabs"
 import AppUrls from "@/utils/urls"
 
 Vue.use(VueCompositionApi)
